feat: Update item and record entities to use float for prices and subtotals
- Changed 'selling_price' in Item entity from integer to float.
- Changed 'subtotal' in Record entity from integer to float.
- Added default value 'unidad' for 'measure_unit' in items table migration.
- Updated ContactValidator to enforce type validation with error messages.
- Enhanced UserRules for email uniqueness and password validation.
- Improved form validation and user experience in various views (Contact, Item, Category, User).
- Refactored show and new views for Business and User to improve layout and validation feedback.
- Added new background images for login and navigation bar.

+++ app/Entities/Item.php
@@ -15,7 +15,7 @@ class Item extends Entity
         'name' => 'string',
         'type' => 'string',
         'cost' => 'float',
-        'selling_price' => 'integer',
+        'selling_price' => 'float',
         'stock' => 'integer',
         'min_stock' => 'integer',
         'measure_unit' => 'string',
